Small Python restructuring
