+++ transforms/universal/fdedup_multi_step/python/src/fdedup/transforms/python/fdedup_bucket_processor_transform.py
@@ -135,20 +135,11 @@ class FdedupBucketProcessorRuntime(DefaultPythonTransformRuntime):
         :param files - list of files to process
         :return: dictionary of transform init params
         """
-        data_access = data_access_factory.create_data_access()
-        # create support classes
-        snapshot_path = self.params.get(minhash_snapshot_directory_key, None)
-        if snapshot_path is None or len(snapshot_path) == 0:
-            mh_path = f"{SnapshotUtils.get_snapshot_folder(data_access)}minhash/minhash_collector_0"
-        else:
-            mh_path = f"{snapshot_path}/minhash_collector_0"
-        self.minhashes = DocsMinHash({"id": 0, "data_access": data_access_factory, "snapshot": mh_path})
-        snapshot_path = self.params.get(doc_id_snapshot_directory_key, None)
-        if snapshot_path is None or len(snapshot_path) == 0:
-            doc_path = f"{SnapshotUtils.get_snapshot_folder(data_access)}docs/doc_collector_0"
-        else:
-            doc_path = f"{snapshot_path}/doc_collector_0"
-        self.doc_collector = DocCollector({"id": 0, "data_access": data_access_factory, "snapshot": doc_path})
+        from fdedup.utils import FdedupSupport
+        self.minhashes = FdedupSupport.createMinhashCurrent(
+            data_access_factory=data_access_factory, directory= self.params.get(minhash_snapshot_directory_key, None))
+        self.doc_collector = FdedupSupport.createDocIDCurrent(
+            data_access_factory=data_access_factory, directory=self.params.get(doc_id_snapshot_directory_key, None))
         self.buckets = BucketsHash({"id": 0, "data_access": data_access_factory, "snapshot": None})
         self.bucket_processor = PythonBucketsHashProcessor({"threshold": self.threshold,
                                                             "docs_collector": self.doc_collector,
