+++ transforms/universal/fdedup_multi_step/python/src/fdedup/utils/__init__.py
@@ -1,13 +1,9 @@
 from fdedup.utils.compute_shingles import compute_shingles
-from fdedup.utils.support_classes import (fuzzy_optimal_param,
-                                          jaccard_distance,
-                                          process_buckets_locally,
-                                          merge_doc_ids,
-                                          remove_docs_buckets,
-                                          NO_SIMILARITY,
-                                          MurmurMH,
-                                          DocCollector,
-                                          DocsMinHash,
-                                          BucketsHash,
-                                          BucketsHashProcessor
-                                          )
+from fdedup.utils.fdedup_classes import (NO_SIMILARITY,
+                                         MurmurMH,
+                                         DocCollector,
+                                         DocsMinHash,
+                                         BucketsHash,
+                                         BucketsHashProcessor
+                                         )
+from fdedup.utils.fdedupsupport import  FdedupSupport
