+++ transforms/universal/fdedup_multi_step/ray/.dockerignore
@@ -1,0 +1,1 @@
+venv/
