+++ transforms/universal/fdedup_multi_step/ray/Dockerfile
@@ -1,0 +1,47 @@
+ARG BASE_IMAGE=docker.io/rayproject/ray:2.24.0-py310
+
+FROM ${BASE_IMAGE}
+
+RUN pip install --upgrade --no-cache-dir pip 
+
+# install pytest
+RUN pip install --no-cache-dir pytest
+
+# Copy and install data processing libraries 
+# These are expected to be placed in the docker context before this is run (see the make image).
+COPY --chown=ray:users data-processing-lib-python/ data-processing-lib-python/
+RUN cd data-processing-lib-python && pip install --no-cache-dir -e .
+COPY --chown=ray:users data-processing-lib-ray/ data-processing-lib-ray/ 
+RUN cd data-processing-lib-ray    && pip install --no-cache-dir -e .
+COPY --chown=ray:users python-transform/  python-transform/
+RUN cd python-transform && pip install --no-cache-dir -e .
+
+# Install ray project source
+COPY --chown=ray:users src/ src/
+COPY --chown=ray:users pyproject.toml pyproject.toml
+COPY --chown=ray:users README.md README.md
+COPY --chown=ray:users images/ images/
+RUN pip install --no-cache-dir -e .
+
+# copy the main() entry point to the image 
+COPY ./src/fdedup_ray/transforms/fdedup_preprocessor_transform_ray.py .
+COPY ./src/fdedup_ray/transforms/fdedup_bucket_processor_transform_ray.py .
+COPY ./src/fdedup_ray/transforms/fdedup_filter_transform_ray.py .
+
+# copy some of the samples in
+COPY ./src/fdedup_ray/examples/fdedup_preprocessor_local_ray.py local/
+COPY ./src/fdedup/examples/fdedup_bucket_processor_local_ray.py local/
+COPY ./src/fdedup/examples/fdedup_filter_local_ray.py local/
+
+# copy test
+COPY test/ test/
+COPY test-data/ test-data/
+
+# Set environment
+ENV PYTHONPATH /home/ray
+
+# Put these at the end since they seem to upset the docker cache.
+ARG BUILD_DATE
+ARG GIT_COMMIT
+LABEL build-date=$BUILD_DATE
+LABEL git-commit=$GIT_COMMIT
